refactor: make chatter list its own widget
This makes the Chatter List widget/window/popup into an actual
BaseWindow instead of a QDockWidget. the QDockWidget only had
functionality on Linux, and it barely functioned since we're not using
it with any dockable layouts/windows
